+++ dev/src/views/ui/attachment.vue
@@ -44,12 +44,13 @@
       :valueFormatter="valueFormatter"
       :uploader="uploader"
       v-model="files"
-      :auto-upload="true"></XAttachment>
+      :auto-upload="true"
+      style="border: 1px red solid; display: inline-block"></XAttachment>
   </div>
 </template>
 <script lang="ts" setup>
   import { ref } from 'vue';
-  import { XAttachment, request, type AttachmentFile, delay } from '@vtj/web';
+  import { XAttachment, type AttachmentFile, delay } from '@vtj/web';
 
   const fileList = ref<AttachmentFile[]>([
     {