+++ dev/src/views/ui/query-form.vue
@@ -1,13 +1,16 @@
 <template>
   <div>
     <XQueryForm ref="formRef" size="default" :model="model" :items="items">
+      <template #custom>
+        <XField label="自定义"></XField>
+      </template>
     </XQueryForm>
     <XAction @click="onClick" label="提交"></XAction>
   </div>
 </template>
 <script lang="ts" setup>
   import { reactive, ref } from 'vue';
-  import { XQueryForm, XAction } from '@vtj/web';
+  import { XQueryForm, XAction, XField, type QueryFormItems } from '@vtj/web';
 
   const formRef = ref();
 
@@ -15,7 +18,7 @@
     F1: 'abc'
   });
 
-  const items = [
+  const items: QueryFormItems = [
     {
       label: '姓名',
       name: 'name',
@@ -27,7 +30,8 @@
           value: 1
         }
       ]
-    }
+    },
+    'custom'
   ];
 
   const onClick = () => {