+++ dev/src/views/ui/verify.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    <XVerify></XVerify>
+  </div>
+</template>
+<script lang="ts" setup>
+  import { XVerify } from '@vtj/web';
+</script>