+++ packages/ui/src/components/query-form/QueryForm.vue
@@ -13,10 +13,10 @@
       :class="collapsedClass"
       :style="collapsedStyle">
       <slot>
-        <XField
-          v-for="item in props.items"
-          v-bind="item"
-          :key="`field_${item.name}`"></XField>
+        <template v-for="item in props.items">
+          <slot v-if="isSlotField(item)" :name="item"></slot>
+          <XField v-else v-bind="item" :key="`field_${item.name}`"></XField>
+        </template>
       </slot>
     </div>
     <template #action>
@@ -43,7 +43,7 @@
   import { ref } from 'vue';
   import { XForm, XAction, XField } from '../';
   import { CaretBottom, CaretTop } from '@vtj/icons';
-  import type { QueryFormEmits } from './types';
+  import type { QueryFormEmits, QueryFormItem } from './types';
   import { queryFormProps } from './props';
   import { useCollapsed } from './hooks';
 
@@ -74,10 +74,20 @@
     formRef.value?.reset(fields);
   };
 
+  const isSlotField = (item: QueryFormItem) => {
+    return typeof item === 'string';
+  };
+
+  const $vtjDynamicSlots = () => {
+    const items = props.items || [];
+    return items.map((n) => isSlotField(n));
+  };
+
   defineExpose({
     validate,
     submit,
     reset,
-    formRef
+    formRef,
+    $vtjDynamicSlots
   });
 </script>