+++ packages/ui/src/components/query-form/types.ts
@@ -2,7 +2,9 @@ import type { ComponentPropsType } from '../shared';
 import { queryFormProps } from './props';
 import type { FieldProps } from '../';
 
-export type QueryFormItems = FieldProps[];
+export type QueryFormItem = string | FieldProps;
+
+export type QueryFormItems = QueryFormItem[];
 
 export type QueryFormProps = ComponentPropsType<typeof queryFormProps>;