+++ packages/ui/src/components/sms-code-input/index.ts
@@ -1,3 +1,0 @@
-import XSmsCodeInput from './SmsCodeInput.vue';
-export { XSmsCodeInput };
-export * from './types';