+++ packages/ui/src/components/verify/index.ts
@@ -1,0 +1,3 @@
+import XVerify from './Verify.vue';
+export { XVerify };
+export * from './types';