+++ packages/ui/src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useIcon';
 export * from './useDisabled';
 export * from './useLoader';
 export * from './useDefer';
+export * from './useDefineSlots';