+++ packages/ui/src/hooks/useDefineSlots.ts
@@ -1,0 +1,5 @@
+import { useSlots } from 'vue';
+export function useDefineSlots(excludes: string[] = []) {
+  const slots = useSlots();
+  return Object.keys(slots).filter((n) => !excludes.includes(n));
+}